+++ android/src/main/java/com/theoplayer/util/TypeUtils.kt
@@ -6,12 +6,12 @@ const val POS_INF_VALUE: Double = -2.0
 object TypeUtils {
   // Make sure we do not send INF or NaN double values over the bridge. It will break debug sessions.
   fun encodeInfNan(v: Double): Double {
-    return when (v) {
-      Double.NaN -> NAN_VALUE
-      Double.POSITIVE_INFINITY -> POS_INF_VALUE
-      else -> {
-        return v
-      }
+    if (v.isNaN()) {
+      return NAN_VALUE
     }
+    if (v.isInfinite()) {
+      return POS_INF_VALUE
+    }
+    return v
   }
 }