+++ ios/THEOplayerRCTView.swift
@@ -7,7 +7,7 @@ import THEOplayerSDK
 public class THEOplayerRCTView: UIView {
     // MARK: Members
     public private(set) var player: THEOplayer?
-    var mainEventHandler: THEOplayerRCTMainEventHandler
+    public private(set) var mainEventHandler: THEOplayerRCTMainEventHandler
     var textTrackEventHandler: THEOplayerRCTTextTrackEventHandler
     var mediaTrackEventHandler: THEOplayerRCTMediaTrackEventHandler
     var adEventHandler: THEOplayerRCTAdsEventHandler
@@ -136,6 +136,8 @@ public class THEOplayerRCTView: UIView {
         self.remoteCommandsManager.destroy()
         self.pipControlsManager.destroy()
         
+        self.destroyBackgroundAudio()
+        self.player?.removeAllIntegrations()
         self.player?.destroy()
         self.player = nil
         if DEBUG_THEOPLAYER_INTERACTION { print("[NATIVE] THEOplayer instance destroyed.") }