+++ ios/backgroundAudio/THEOplayerRCTView+BackgroundAudioConfig.swift
@@ -11,6 +11,7 @@ struct BackgroundAudioConfig {
 
 extension THEOplayerRCTView {
     func initBackgroundAudio() {}
+    func destroyBackgroundAudio() {}
     public func shouldContinueAudioPlaybackInBackground() -> Bool { return false }
 }
 
@@ -22,12 +23,23 @@ extension THEOplayerRCTView: BackgroundPlaybackDelegate {
         self.player?.backgroundPlaybackDelegate = self
     }
     
+    func destroyBackgroundAudio() {
+        guard let player = self.player else {
+            return
+        }
+        player.backgroundPlaybackDelegate = DefaultBackgroundPlaybackDelegate()
+    }
+    
     public func shouldContinueAudioPlaybackInBackground() -> Bool {
         // Make sure to go to the background with updated NowPlayingInfo
         self.nowPlayingManager.updateNowPlaying()
         
         return self.backgroundAudioConfig.enabled
     }
+}
+
+struct DefaultBackgroundPlaybackDelegate: BackgroundPlaybackDelegate {
+    func shouldContinueAudioPlaybackInBackground() -> Bool { false }
 }
 
 #endif